feat(105988): cria arquivo helpers e move algumas funções pra lá
