+++ src/components/screens/PreRecebimento/FichaTecnica/components/Cadastrar/index.tsx
@@ -2,15 +2,7 @@ import React, { useContext, useEffect, useRef, useState } from "react";
 import "./styles.scss";
 import { Field, Form } from "react-final-form";
 import Label from "components/Shareable/Label";
-import {
-  getListaCompletaProdutosLogistica,
-  getNomesMarcas,
-  getNomesFabricantes,
-  getInformacoesNutricionaisOrdenadas,
-} from "services/produto.service";
-import { getEnderecoPorCEP } from "services/cep.service";
-import { getTerceirizadaUUID } from "services/terceirizada.service";
-import { getUnidadesDeMedidaLogistica } from "services/cronograma.service";
+import { getInformacoesNutricionaisOrdenadas } from "services/produto.service";
 import {
   required,
   email,
@@ -24,7 +16,6 @@ import MaskedInputText from "components/Shareable/Input/MaskedInputText";
 
 import Collapse, { CollapseControl } from "components/Shareable/Collapse";
 import MeusDadosContext from "context/MeusDadosContext";
-import createDecorator from "final-form-calculate";
 import {
   BUTTON_TYPE,
   BUTTON_STYLE,
@@ -36,38 +27,45 @@ import {
   editaRascunhoFichaTecnica,
   getFichaTecnica,
 } from "services/fichaTecnica.service";
-import { exibeError, removeCaracteresEspeciais } from "helpers/utilities";
-import {
-  toastError,
-  toastSuccess,
-} from "../../../../../Shareable/Toast/dialogs";
-import { getListaFiltradaAutoCompleteSelect } from "../../../../../../helpers/autoCompleteSelect";
+import { exibeError } from "helpers/utilities";
+import { toastError, toastSuccess } from "components/Shareable/Toast/dialogs";
+import { getListaFiltradaAutoCompleteSelect } from "helpers/autoCompleteSelect";
 import AutoCompleteSelectField from "components/Shareable/AutoCompleteSelectField";
 import FormPereciveis from "./components/FormPereciveis";
 import FormNaoPereciveis from "./components/FormNaoPereciveis";
 import { OnChange } from "react-final-form-listeners";
 import TabelaNutricional from "components/Shareable/TabelaNutricional";
 import Select from "components/Shareable/Select";
-
 import ModalCadastrarItemIndividual from "components/Shareable/ModalCadastrarItemIndividual";
 
+import { MeusDadosInterfaceOuter } from "context/MeusDadosContext/interfaces";
+
 import {
-  CategoriaFichaTecnicaChoices,
   FichaTecnicaDetalhada,
   OptionsGenerico,
 } from "interfaces/pre_recebimento.interface";
-import { MeusDadosInterfaceOuter } from "context/MeusDadosContext/interfaces";
 import { TerceirizadaComEnderecoInterface } from "interfaces/terceirizada.interface";
 import {
   ResponseFichaTecnicaDetalhada,
   ResponseInformacoesNutricionais,
 } from "interfaces/responses.interface";
 import { InformacaoNutricional } from "interfaces/produto.interface";
-import {
-  FichaTecnicaPayload,
-  InformacoesNutricionaisFichaTecnicaPayload,
-} from "../../interfaces";
+
 import InfoAcondicionamentoPereciveis from "./components/InfoAcondicionamentoPereciveis";
+
+import { FichaTecnicaPayload } from "../../interfaces";
+import {
+  carregarFabricantes,
+  carregarMarcas,
+  carregarProdutos,
+  carregarTerceirizada,
+  carregarUnidadesMedida,
+  cepCalculator,
+  formataPayload,
+  geraInitialValues,
+  validaProximo,
+  validaRascunho,
+} from "../../helpers";
 
 const ITENS_STEPS = [
   {
@@ -101,7 +99,9 @@ export default () => {
   const [ficha, setFicha] = useState<FichaTecnicaDetalhada>(
     {} as FichaTecnicaDetalhada
   );
-  const [initialValues, setInitialValues] = useState<FichaTecnicaPayload>({});
+  const [initialValues, setInitialValues] = useState<FichaTecnicaPayload>(
+    {} as FichaTecnicaPayload
+  );
   const [stepAtual, setStepAtual] = useState(0);
   const listaCompletaInformacoesNutricionais = useRef<InformacaoNutricional[]>(
     []
@@ -114,43 +114,11 @@ export default () => {
 
   const onSubmit = (): void => {};
 
-  const carregarProdutos = async () => {
-    const response = await getListaCompletaProdutosLogistica();
-    setProdutosOptions(response.data.results);
-  };
-
-  const carregarMarcas = async () => {
-    const response = await getNomesMarcas();
-    setMarcasOptions(response.data.results);
-  };
-
-  const carregarFabricantes = async () => {
-    const response = await getNomesFabricantes();
-    setFabricantesOptions(response.data.results);
-  };
-
-  const carregarUnidadesMedida = async () => {
-    const response = await getUnidadesDeMedidaLogistica();
-    setUnidadesMedidaOptions(response.data.results);
-  };
-
-  const carregarTerceirizada = async () => {
-    if (ficha.empresa?.uuid) {
-      const response = await getTerceirizadaUUID(ficha.empresa.uuid);
-      setProponente(response.data);
-    } else if (meusDados) {
-      const response = await getTerceirizadaUUID(
-        meusDados.vinculo_atual.instituicao.uuid
-      );
-      setProponente(response.data);
-    }
-  };
-
   const atualizarDadosCarregados = async () => {
     setCarregando(true);
-    await carregarProdutos();
-    await carregarMarcas();
-    await carregarFabricantes();
+    await carregarProdutos(setProdutosOptions);
+    await carregarMarcas(setMarcasOptions);
+    await carregarFabricantes(setFabricantesOptions);
     setCarregando(false);
   };
 
@@ -159,170 +127,13 @@ export default () => {
     setShowModal(true);
   };
 
-  const cepCalculator = createDecorator({
-    field: "cep_fabricante",
-    updates: {
-      dummy: (minimumValue, allValues) => buscaCEP(minimumValue, allValues),
-    },
-  });
-
-  const buscaCEP = async (cep: string, values: FichaTecnicaPayload) => {
-    if (cep?.length === 9) {
-      const response = await getEnderecoPorCEP(cep);
-      if (response.status === 200 && !response.data.erro) {
-        const { data } = response;
-        values.bairro_fabricante = data.bairro;
-        values.cidade_fabricante = data.localidade;
-        values.endereco_fabricante = data.logradouro;
-        values.estado_fabricante = data.uf;
-        setDesabilitaEndereco(true);
-      } else {
-        setDesabilitaEndereco(false);
-      }
-    }
-  };
-
-  const geraInitialValues = (ficha: FichaTecnicaDetalhada) => {
-    const valuesInformacoesNutricionais = {};
-    ficha?.informacoes_nutricionais.forEach((informacao) => {
-      valuesInformacoesNutricionais[
-        `quantidade_por_100g_${informacao.informacao_nutricional.uuid}`
-      ] = informacao.quantidade_por_100g;
-      valuesInformacoesNutricionais[
-        `quantidade_porcao_${informacao.informacao_nutricional.uuid}`
-      ] = informacao.quantidade_porcao;
-      valuesInformacoesNutricionais[
-        `valor_diario_${informacao.informacao_nutricional.uuid}`
-      ] = informacao.valor_diario;
-    });
-
-    const valuesSelect = {};
-    ficha?.informacoes_nutricionais
-      .filter(({ informacao_nutricional }) => !informacao_nutricional.eh_fixo)
-      .forEach((informacao, index) => {
-        valuesSelect[`informacao_adicional_${index}`] =
-          informacao.informacao_nutricional.uuid;
-      });
-
-    return {
-      ...initialValues,
-      produto: ficha.produto?.nome,
-      marca: ficha.marca.uuid,
-      categoria: ficha.categoria as CategoriaFichaTecnicaChoices,
-      pregao_chamada_publica: ficha.pregao_chamada_publica,
-      fabricante: ficha.fabricante?.nome,
-      cnpj_fabricante: ficha.cnpj_fabricante,
-      cep_fabricante: ficha.cep_fabricante,
-      endereco_fabricante: ficha.endereco_fabricante,
-      numero_fabricante: ficha.numero_fabricante,
-      complemento_fabricante: ficha.complemento_fabricante,
-      bairro_fabricante: ficha.bairro_fabricante,
-      cidade_fabricante: ficha.cidade_fabricante,
-      estado_fabricante: ficha.estado_fabricante,
-      email_fabricante: ficha.email_fabricante,
-      telefone_fabricante: ficha.telefone_fabricante,
-      prazo_validade: ficha.prazo_validade,
-      numero_registro: ficha.numero_registro,
-      agroecologico: booleanToString(ficha.agroecologico),
-      organico: booleanToString(ficha.organico),
-      mecanismo_controle: ficha.mecanismo_controle,
-      componentes_produto: ficha.componentes_produto,
-      alergenicos: booleanToString(ficha.alergenicos),
-      ingredientes_alergenicos: ficha.ingredientes_alergenicos,
-      gluten: booleanToString(ficha.gluten),
-      lactose: booleanToString(ficha.lactose),
-      lactose_detalhe: ficha.lactose_detalhe,
-      porcao: ficha.porcao,
-      unidade_medida_porcao: ficha.unidade_medida_porcao?.uuid,
-      valor_unidade_caseira: ficha.valor_unidade_caseira,
-      unidade_medida_caseira: ficha.unidade_medida_caseira,
-      ...valuesInformacoesNutricionais,
-      ...valuesSelect,
-    };
-  };
-
-  const formataPayload = (values: Record<string, any>): FichaTecnicaPayload => {
-    let payload: FichaTecnicaPayload = {
-      produto: produtosOptions.find((p) => p.nome === values.produto)?.uuid,
-      marca: values.marca,
-      categoria: values.categoria,
-      pregao_chamada_publica: values.pregao_chamada_publica,
-      empresa: proponente.uuid,
-      fabricante:
-        fabricantesOptions.find((p) => p.nome === values.fabricante)?.uuid ||
-        null,
-      cnpj_fabricante: removeCaracteresEspeciais(values.cnpj_fabricante) || "",
-      cep_fabricante: removeCaracteresEspeciais(values.cep_fabricante) || "",
-      endereco_fabricante: values.endereco_fabricante || "",
-      numero_fabricante: values.numero_fabricante || "",
-      complemento_fabricante: values.complemento_fabricante || "",
-      bairro_fabricante: values.bairro_fabricante || "",
-      cidade_fabricante: values.cidade_fabricante || "",
-      estado_fabricante: values.estado_fabricante || "",
-      email_fabricante: values.email_fabricante || "",
-      telefone_fabricante:
-        removeCaracteresEspeciais(values.telefone_fabricante) || "",
-      prazo_validade: values.prazo_validade || "",
-      componentes_produto: values.componentes_produto || "",
-      alergenicos: stringToBoolean(values.alergenicos as string),
-      gluten: stringToBoolean(values.gluten as string),
-      lactose: stringToBoolean(values.lactose as string),
-      mecanismo_controle: "",
-      ingredientes_alergenicos: "",
-      lactose_detalhe: "",
-      numero_registro: "",
-      porcao: values.porcao || "",
-      unidade_medida_porcao: values.unidade_medida_porcao || null,
-      valor_unidade_caseira: values.valor_unidade_caseira || "",
-      unidade_medida_caseira: values.unidade_medida_caseira || "",
-      informacoes_nutricionais: formataInformacoesNutricionais(values),
-    };
-    if (payload.alergenicos) {
-      payload.ingredientes_alergenicos = values.ingredientes_alergenicos || "";
-    }
-    if (payload.lactose) {
-      payload.lactose_detalhe = values.lactose_detalhe || "";
-    }
-    if (payload.categoria === "PERECIVEIS") {
-      payload = {
-        ...payload,
-        numero_registro: values.numero_registro || "",
-        agroecologico: stringToBoolean(values.agroecologico as string),
-        organico: stringToBoolean(values.organico as string),
-      };
-      if (payload.organico) {
-        payload.mecanismo_controle = values.mecanismo_controle || "";
-      }
-    }
-
-    return payload;
-  };
-
-  const stringToBoolean = (str: string): boolean =>
-    str === "1" ? true : str === "0" ? false : undefined;
-
-  const booleanToString = (str: boolean): string =>
-    str === true ? "1" : str === false ? "0" : undefined;
-
-  const formataInformacoesNutricionais = (values: Record<string, any>) => {
-    const uuids_informacoes = Object.keys(values)
-      .filter((key) => key.startsWith("quantidade_por_100g_"))
-      .map((key) => key.split("_").pop());
-
-    const payload = uuids_informacoes.map((uuid) => {
-      return {
-        informacao_nutricional: uuid,
-        quantidade_por_100g: values[`quantidade_por_100g_${uuid}`],
-        quantidade_porcao: values[`quantidade_porcao_${uuid}`],
-        valor_diario: values[`valor_diario_${uuid}`],
-      };
-    });
-
-    return payload as InformacoesNutricionaisFichaTecnicaPayload[];
-  };
-
   const salvarRascunho = async (values: FichaTecnicaPayload) => {
-    const payload = formataPayload(values);
+    const payload = formataPayload(
+      values,
+      proponente,
+      produtosOptions,
+      fabricantesOptions
+    );
 
     try {
       setCarregando(true);
@@ -352,8 +163,6 @@ export default () => {
     listaCompletaInformacoesNutricionais.current =
       responseInformacoes.data.results;
 
-    let initialValues = {};
-
     const urlParams = new URLSearchParams(window.location.search);
     const uuid = urlParams.get("uuid");
     if (uuid) {
@@ -366,20 +175,17 @@ export default () => {
         );
 
       setFicha(fichaTecnica);
-
-      initialValues = { ...geraInitialValues(fichaTecnica) };
+      setInitialValues(geraInitialValues(fichaTecnica));
     }
-
-    setInitialValues(initialValues);
   };
 
   useEffect(() => {
     (async () => {
       setCarregando(true);
-      await carregarProdutos();
-      await carregarMarcas();
-      await carregarFabricantes();
-      await carregarUnidadesMedida();
+      await carregarProdutos(setProdutosOptions);
+      await carregarMarcas(setMarcasOptions);
+      await carregarFabricantes(setFabricantesOptions);
+      await carregarUnidadesMedida(setUnidadesMedidaOptions);
       await carregarDados();
       setCarregando(false);
     })();
@@ -389,80 +195,11 @@ export default () => {
     (async () => {
       if (!proponente.uuid) {
         setCarregando(true);
-        await carregarTerceirizada();
+        await carregarTerceirizada(ficha, meusDados, setProponente);
         setCarregando(false);
       }
     })();
   }, [meusDados, ficha]);
-
-  const validaRascunho = (values: FichaTecnicaPayload): boolean => {
-    return (
-      !values.produto ||
-      !values.marca ||
-      !values.categoria ||
-      !values.pregao_chamada_publica
-    );
-  };
-
-  const validaProximo = (
-    values: FichaTecnicaPayload,
-    errors: Record<string, string>
-  ): boolean => {
-    const validaStepMap = [
-      validaProximoIdentificacaoProduto,
-      validaProximoInformacoesNutricionais,
-    ];
-
-    return validaStepMap[stepAtual](values, errors);
-  };
-
-  const validaProximoIdentificacaoProduto = (
-    values: FichaTecnicaPayload,
-    errors: Record<string, string>
-  ): boolean => {
-    const campoAlergenicosValido =
-      (values.alergenicos === "1" && values.ingredientes_alergenicos) ||
-      values.alergenicos === "0";
-
-    const campoComLactoseValido =
-      (values.lactose === "1" && values.lactose_detalhe) ||
-      values.lactose === "0";
-
-    const campoOrganicoValido =
-      (values.organico === "1" && values.mecanismo_controle) ||
-      values.organico === "0";
-
-    const camposFormPereciveisValidos =
-      values.numero_registro && values.agroecologico && campoOrganicoValido;
-
-    return (
-      Object.keys(errors).length !== 0 ||
-      !values.produto ||
-      !values.marca ||
-      !values.categoria ||
-      !values.pregao_chamada_publica ||
-      !values.fabricante ||
-      !values.prazo_validade ||
-      !values.componentes_produto ||
-      !values.gluten ||
-      !campoAlergenicosValido ||
-      !campoComLactoseValido ||
-      (values.categoria === "PERECIVEIS" && !camposFormPereciveisValidos)
-    );
-  };
-
-  const validaProximoInformacoesNutricionais = (
-    values: FichaTecnicaPayload,
-    errors: Record<string, string>
-  ): boolean => {
-    return (
-      Object.keys(errors).length !== 0 ||
-      !values.porcao ||
-      !values.unidade_medida_porcao ||
-      !values.valor_unidade_caseira ||
-      !values.unidade_medida_caseira
-    );
-  };
 
   return (
     <Spin tip="Carregando..." spinning={carregando}>
@@ -471,7 +208,7 @@ export default () => {
           <Form
             onSubmit={onSubmit}
             initialValues={initialValues}
-            decorators={[cepCalculator]}
+            decorators={[cepCalculator(setDesabilitaEndereco)]}
             render={({ form, handleSubmit, values, errors }) => (
               <form onSubmit={handleSubmit}>
                 <div className="steps">
@@ -944,7 +681,11 @@ export default () => {
                       style={BUTTON_STYLE.GREEN_OUTLINE}
                       className="float-end ms-3"
                       onClick={() => setStepAtual((stepAtual) => stepAtual + 1)}
-                      disabled={validaProximo(values, errors)}
+                      disabled={validaProximo(
+                        values as FichaTecnicaPayload,
+                        errors,
+                        stepAtual
+                      )}
                     />
                   </div>
                 )}
@@ -955,8 +696,10 @@ export default () => {
                     type={BUTTON_TYPE.BUTTON}
                     style={BUTTON_STYLE.GREEN_OUTLINE}
                     className="float-end ms-3"
-                    onClick={() => salvarRascunho(values)}
-                    disabled={validaRascunho(values)}
+                    onClick={() =>
+                      salvarRascunho(values as FichaTecnicaPayload)
+                    }
+                    disabled={validaRascunho(values as FichaTecnicaPayload)}
                   />
                 </div>
